+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/ItemMixin.java
@@ -1,6 +1,6 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
-import insane96mcp.iguanatweaksreborn.module.hungerhealth.feature.FoodConsuming;
+import insane96mcp.iguanatweaksreborn.module.hungerhealth.feature.Food;
 import insane96mcp.insanelib.base.Feature;
 import net.minecraft.world.item.BowlFoodItem;
 import net.minecraft.world.item.Item;
@@ -15,17 +15,17 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public class ItemMixin {
 	@Inject(at = @At("HEAD"), method = "getUseDuration", cancellable = true)
 	public void getUseDuration(ItemStack stack, CallbackInfoReturnable<Integer> callbackInfo) {
-		if (!FoodConsuming.eatingSpeedBasedOffFood || !Feature.isEnabled(FoodConsuming.class))
+		if (!Food.eatingSpeedBasedOffFood || !Feature.isEnabled(Food.class))
 			return;
 
 		if (stack.getItem().isEdible()) {
-			callbackInfo.setReturnValue(FoodConsuming.getFoodConsumingTime(stack));
+			callbackInfo.setReturnValue(Food.getFoodConsumingTime(stack));
 		}
 	}
 
 	@Inject(at = @At("RETURN"), method = "getUseAnimation", cancellable = true)
 	public void getUseAnimation(ItemStack stack, CallbackInfoReturnable<UseAnim> callbackInfo) {
-		if (!Feature.isEnabled(FoodConsuming.class))
+		if (!Feature.isEnabled(Food.class))
 			return;
 
 		if (stack.getItem() instanceof BowlFoodItem) {