+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/PotionItemMixin.java
@@ -1,6 +1,6 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
-import insane96mcp.iguanatweaksreborn.module.hungerhealth.feature.FoodConsuming;
+import insane96mcp.iguanatweaksreborn.module.hungerhealth.feature.Food;
 import insane96mcp.insanelib.base.Feature;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.PotionItem;
@@ -13,7 +13,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public class PotionItemMixin {
 	@Inject(at = @At("RETURN"), method = "getUseDuration", cancellable = true)
 	public void getUseDuration(ItemStack stack, CallbackInfoReturnable<Integer> callbackInfo) {
-		if (!FoodConsuming.fasterPotionConsuming || !Feature.isEnabled(FoodConsuming.class))
+		if (!Food.fasterPotionConsuming || !Feature.isEnabled(Food.class))
 			return;
 
 		callbackInfo.setReturnValue(20);